% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA.R
\name{aba_array_dims}
\alias{aba_array_dims}
\title{Title}
\usage{
aba_array_dims(resolution = 200)
}
\arguments{
\item{resolution}{Resolution of the returned array dimensions ("200um" or "25um")}
}
\value{

}
\description{
Title
}
