% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA.R
\name{ccf_annotation}
\alias{ccf_annotation}
\title{Get CCFv3 Annotation}
\usage{
ccf_annotation(age = "Adult", make_array = FALSE, array_spacing = 200)
}
\arguments{
\item{age}{One of c("Adult","E11","E13","E15","E18","P4","P14","P28","P56")}

\item{make_array}{Logical argument whether to return the result as a 3D array instead of an object of class 'sf'.}

\item{array_spacing}{numerical value indicating spacing between array point (usually 25 or 200) in uM.}
}
\value{

}
\description{
Get CCFv3 Annotation
}
