% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coordinates,SpatialExperiment-method}
\alias{coordinates,SpatialExperiment-method}
\title{Generic to access SpatialExperiment coordinates from spatialMap slot}
\usage{
\S4method{coordinates}{SpatialExperiment}(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment coordinates from spatialMap slot
}
