% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coordinates}
\alias{coordinates}
\title{Generic to access SpatialExperiment coordinates from spatialMapslot}
\usage{
coordinates(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment coordinates from spatialMapslot
}
