% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{embeddings<-,SpatialExperiment-method}
\alias{embeddings<-,SpatialExperiment-method}
\title{Method to add to SpatialExperiment embedding slot list.}
\usage{
\S4method{embeddings}{SpatialExperiment}(x, name) <- value
}
\arguments{
\item{x}{A SpatialExperiment object.}

\item{value}{A LinearEmbeddingObject to add to embeddings(x).}

\item{name}{Name of the LEM object to add}
}
\description{
Method to add to SpatialExperiment embedding slot list.
}
