% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{embeddings<-}
\alias{embeddings<-}
\title{Generic to set SpatialExperiment embedding slot element}
\usage{
embeddings(x, name) <- value
}
\arguments{
\item{x}{A SpatialExperiment object.}

\item{value}{A LinearEmbeddingObject to add to embeddings(x).}

\item{name}{Name of the LEM object to add.}
}
\description{
Generic to set SpatialExperiment embedding slot element
}
