% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{embeddings,SpatialExperiment-method}
\alias{embeddings,SpatialExperiment-method}
\title{Generic to access SpatialExperiment embedding slot}
\usage{
\S4method{embeddings}{SpatialExperiment}(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment embedding slot
}
