% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{embeddings}
\alias{embeddings}
\title{Generic to access SpatialExperiment embedding slot}
\usage{
embeddings(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment embedding slot
}
