% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{exprs,SpatialExperiment-method}
\alias{exprs,SpatialExperiment-method}
\title{Method to access SpatialExperiment count matrix}
\usage{
\S4method{exprs}{SpatialExperiment}(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Method to access SpatialExperiment count matrix
}
