% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{exprs}
\alias{exprs}
\title{Generic to access SpatialExperiment count matrix}
\usage{
exprs(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment count matrix
}
