% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData<-,SpatialExperiment-method}
\alias{fData<-,SpatialExperiment-method}
\title{Method to set SpatialExperiment rowData table}
\usage{
\S4method{fData}{SpatialExperiment}(x) <- value
}
\arguments{
\item{x}{A SpatialExperiment object.}

\item{value}{A data frame to set to colData table.}
}
\description{
Method to set SpatialExperiment rowData table
}
