% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData,SpatialExperiment-method}
\alias{fData,SpatialExperiment-method}
\title{Generic to access SpatialExperiment rowData table}
\usage{
\S4method{fData}{SpatialExperiment}(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment rowData table
}
