% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fData}
\alias{fData}
\title{Generic to access SpatialExperiment rowData table}
\usage{
fData(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment rowData table
}
