% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA.R
\name{fetch_current_sectionData_objects}
\alias{fetch_current_sectionData_objects}
\title{fetch_current_sectionData_objects}
\usage{
fetch_current_sectionData_objects(age = "P56")
}
\arguments{
\item{age}{}
}
\value{

}
\description{
fetch_current_sectionData_objects
}
