% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABA.R
\name{get_aba_ish_raw}
\alias{get_aba_ish_raw}
\title{Retrieve raw values for a particular SectionData object from ABA.
Very much adapted from github.com/alleninstitute/cocoframer}
\usage{
get_aba_ish_raw(api_id, values = "energy")
}
\arguments{
\item{values}{}
}
\value{

}
\description{
Retrieve raw values for a particular SectionData object from ABA.
Very much adapted from github.com/alleninstitute/cocoframer
}
