% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_processing.R
\name{image_reg_visium}
\alias{image_reg_visium}
\title{Image registration to 10x Visium fiducials}
\usage{
image_reg_visium(img)
}
\arguments{
\item{img}{image pointer}
}
\value{

}
\description{
Image registration to 10x Visium fiducials
}
\examples{

}
