% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myColorRamp}
\alias{myColorRamp}
\title{myColoRamp utility}
\usage{
myColorRamp(values, palette = viridis(255))
}
\arguments{
\item{values}{A vector of expression values to map to palette}

\item{palette}{A color palette}
}
\value{

}
\description{
myColoRamp utility
}
