% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{NewSpatialExperiment}
\alias{NewSpatialExperiment}
\title{Create a new SpatialExperiment object.}
\usage{
NewSpatialExperiment(
  expression_data,
  spatial_coords,
  barcode_metadata = NULL,
  gene_metadata = NULL
)
}
\arguments{
\item{expression_data}{expression data matrix for an experiment, can be a
sparseMatrix.}

\item{barcode_metadata}{data frame containing attributes of individual pixels/voxels/barcodes,
where \code{row.names(barcode_metadata) = colnames(expression_data)}.}

\item{gene_metadata}{data frame containing attributes of features
(e.g. genes), where
\code{row.names(gene_metadata) = row.names(expression_data)}.}
}
\value{
a new SpatialExperiment object
}
\description{
Create a new SpatialExperiment object.
}
\examples{


}
