% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pData<-,SpatialExperiment-method}
\alias{pData<-,SpatialExperiment-method}
\title{Method to set SpatialExperiment colData table}
\usage{
\S4method{pData}{SpatialExperiment}(x) <- value
}
\arguments{
\item{x}{A SpatialExperiment object.}

\item{value}{A data frame to set to colData table.}
}
\description{
Method to set SpatialExperiment colData table
}
