% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pData<-}
\alias{pData<-}
\title{Generic to set SpatialExperiment colData table}
\usage{
pData(x) <- value
}
\arguments{
\item{x}{A SpatialExperiment object.}

\item{value}{A data frame to set to colData table.}
}
\description{
Generic to set SpatialExperiment colData table
}
