% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pData,SpatialExperiment-method}
\alias{pData,SpatialExperiment-method}
\title{Method to access SpatialExperiment colData table}
\usage{
\S4method{pData}{SpatialExperiment}(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Method to access SpatialExperiment colData table
}
