% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pData}
\alias{pData}
\title{Generic to access SpatialExperiment colData table}
\usage{
pData(x)
}
\arguments{
\item{x}{A SpatialExperiment object.}
}
\description{
Generic to access SpatialExperiment colData table
}
