% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimReductions.R
\name{reduce_dimension_spatial}
\alias{reduce_dimension_spatial}
\title{Title}
\usage{
reduce_dimension_spatial(
  se,
  method = c("NNMF"),
  rank = 20,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{se}{A SpatialExperiment object}

\item{method}{Pattern discovery method.  Must be one of c("NNMF").  [default: "NNMF"]}

\item{rank}{Where applicable, the number of dimensions to reduce into (e.g. 'k') [default: 20]}

\item{verbose}{A logical indicating verbose output. [default: FALSE]}

\item{...}{}
}
\value{

}
\description{
Title
}
