% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizations.R
\name{spatial_plot_3d}
\alias{spatial_plot_3d}
\title{3D spatial plot}
\usage{
spatial_plot_3d(se, color_by = NULL, shape_by = NULL, plot_images = FALSE)
}
\arguments{
\item{se}{A SpatialExperiment Object}

\item{color_by}{A column of pData(se) used to color points (default:NULL)}

\item{shape_by}{A column of pData(se) used as shape aesthetic for points (default:NULL)}

\item{plot_images}{A logical value to determine whether or not section images should be plotted (default: FALSE)}
}
\value{

}
\description{
3D spatial plot
}
