% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialExperiment.R
\docType{class}
\name{SpatialExperiment}
\alias{SpatialExperiment}
\alias{SpatialExperiment-class}
\title{The SpatialExperiment class}
\description{
The SpatialExperiment class
}
\section{Fields}{

\describe{
\item{\code{images}}{SimpleList of images associated with each slide or assay}

\item{\code{clusters}}{SimpleList of cluster information for different
dimensionality reduction.}
}}

